Rename TEdgeConfigRepository::load_new as load
