+++ crates/common/tedge_config/src/lib.rs
@@ -5,7 +5,7 @@ pub mod tedge_config_cli;
 pub use self::tedge_config_cli::config_setting::*;
 pub use self::tedge_config_cli::error::*;
 pub use self::tedge_config_cli::models::*;
-pub use self::tedge_config_cli::new::*;
+pub use self::tedge_config_cli::tedge_config::*;
 pub use self::tedge_config_cli::tedge_config_location::*;
 pub use self::tedge_config_cli::tedge_config_repository::*;
 pub use camino::Utf8Path as Path;
@@ -15,5 +15,5 @@ pub use certificate::CertificateError;
 /// loads the new tedge config from system default
 pub fn get_new_tedge_config() -> Result<TEdgeConfig, TEdgeConfigError> {
     let tedge_config_location = TEdgeConfigLocation::default();
-    TEdgeConfigRepository::new(tedge_config_location).load_new()
+    TEdgeConfigRepository::new(tedge_config_location).load()
 }
