+++ crates/core/plugin_sm/tests/plugin.rs
@@ -232,7 +232,7 @@ mod tests {
         let tmpfile = make_config(100)?;
         let config_location =
             TEdgeConfigLocation::from_custom_root(tmpfile.path().to_str().unwrap());
-        let config = tedge_config::TEdgeConfigRepository::new(config_location).load_new()?;
+        let config = tedge_config::TEdgeConfigRepository::new(config_location).load()?;
 
         let plugin = ExternalPluginCommand::new(
             "test",
