+++ crates/core/tedge/src/cli/certificate/cli.rs
@@ -32,7 +32,7 @@ pub enum TEdgeCertCli {
 
 impl BuildCommand for TEdgeCertCli {
     fn build_command(self, context: BuildContext) -> Result<Box<dyn Command>, ConfigError> {
-        let config = context.config_repository.load_new()?;
+        let config = context.config_repository.load()?;
 
         let cmd = match self {
             TEdgeCertCli::Create { id } => {
