+++ crates/core/tedge/src/cli/init.rs
@@ -60,7 +60,7 @@ impl TEdgeInitCmd {
             ),
         )?;
 
-        let config = self.context.config_repository.load_new()?;
+        let config = self.context.config_repository.load()?;
 
         create_directory(
             config.logs.path.join("tedge"),
