+++ crates/core/tedge/src/cli/mqtt/cli.rs
@@ -42,7 +42,7 @@ pub enum TEdgeMqttCli {
 
 impl BuildCommand for TEdgeMqttCli {
     fn build_command(self, context: BuildContext) -> Result<Box<dyn Command>, crate::ConfigError> {
-        let config = context.config_repository.load_new()?;
+        let config = context.config_repository.load()?;
         let auth_config = config.mqtt_client_auth_config();
 
         let cmd = {
