+++ crates/core/tedge_agent/src/restart_manager/config.rs
@@ -18,7 +18,7 @@ impl RestartManagerConfig {
     ) -> Result<RestartManagerConfig, tedge_config::TEdgeConfigError> {
         let config_repository =
             tedge_config::TEdgeConfigRepository::new(tedge_config_location.clone());
-        let tedge_config = config_repository.load_new()?;
+        let tedge_config = config_repository.load()?;
 
         let tmp_dir = tedge_config.tmp.path.clone();
         let config_dir = tedge_config_location.tedge_config_root_path.clone();
