+++ crates/core/tedge_agent/src/software_manager/config.rs
@@ -34,7 +34,7 @@ impl SoftwareManagerConfig {
     ) -> Result<SoftwareManagerConfig, tedge_config::TEdgeConfigError> {
         let config_repository =
             tedge_config::TEdgeConfigRepository::new(tedge_config_location.clone());
-        let tedge_config = config_repository.load_new()?;
+        let tedge_config = config_repository.load()?;
 
         let config_dir = tedge_config_location.tedge_config_root_path.clone();
