+++ crates/core/tedge_mapper/src/core/component.rs
@@ -10,7 +10,7 @@ pub trait TEdgeComponent: Sync + Send {
     fn mqtt_config(&self) -> Result<mqtt_channel::Config, anyhow::Error> {
         let config_repository =
             tedge_config::TEdgeConfigRepository::new(tedge_config::TEdgeConfigLocation::default());
-        let tedge_config = config_repository.load_new()?;
+        let tedge_config = config_repository.load()?;
 
         let mqtt_config = tedge_config
             .mqtt_config()?
