+++ crates/core/tedge_mapper/src/main.rs
@@ -88,8 +88,7 @@ async fn main() -> anyhow::Result<()> {
 
     let tedge_config_location =
         tedge_config::TEdgeConfigLocation::from_custom_root(&mapper_opt.config_dir);
-    let config =
-        tedge_config::TEdgeConfigRepository::new(tedge_config_location.clone()).load_new()?;
+    let config = tedge_config::TEdgeConfigRepository::new(tedge_config_location.clone()).load()?;
 
     let log_level = if mapper_opt.debug {
         tracing::Level::TRACE
